Refactor client synchronizer to model importer
The client synchronizer was a misnomer, as it would only import contacts
as clients. It is aptly renamed to contact importer. The interface now
has a more loose type constraint, from Contact to Model, to reuse the
interface for importing any Moneybird entity.

The event subscriber and webhook event and controller have been updated
to reflect the ability to import any entity, as long as it has an
importer that supports the entity.